+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 - Provide ability within Report Builder to export it to PDF ([#298](https://gitlab.heigit.org/climate-action/web-app/-/issues/298))
 - Users can now (soft)delete computations from the list
+- Simple in-memory cache for `/metadata` responses ([#155](https://gitlab.heigit.org/climate-action/web-app/-/issues/155))
 
 ## [2.5.1](https://gitlab.heigit.org/climate-action/web-app/-/releases/2.5.1) - 2025-07-21